# Conflicts:
#	abgabe/bibliography.bib
#	abgabe/nbody.tex

+++ benchmark/benchmark_nbody_ocl.py
@@ -2,6 +2,9 @@
 import os
 import time
 import sys
+import math
+
+current_milli_time = lambda: int(round(time.time() * 1000))
 
 if (len(sys.argv) != 4):
     print "Missing arguments.\nUsage: ./benchmark_nbody_ocl.py [num_bodies] [simulation_steps] [repeats]"
@@ -11,20 +14,32 @@ num_bodies = int(sys.argv[1])
 simulation_steps = int(sys.argv[2])
 num_repeats = int(sys.argv[3])
 
-# Benchmarks the OpenCL implementation
-current_milli_time = lambda: int(round(time.time() * 1000))
+if not (num_bodies != 0 and ((num_bodies & (num_bodies - 1)) == 0)):
+    print "{} is not a power of 2. Please provide num_bodies that is a power of 2.".format(num_bodies)
+    sys.exit()
 
 path_to_body_generation = "../resources/generateRandomGalaxy.py"
 path_to_simulation_exe = "../cmake-build-debug/nbody_psys17"
+path_to_plot_exe = "./plot_ocl_benchmark.py"
+
+if not os.path.isfile(path_to_body_generation):
+    print "{} does not exist!".format(path_to_body_generation)
+    sys.exit();
+
+if not os.path.isfile(path_to_simulation_exe):
+    print "{} does not exist!".format(path_to_simulation_exe);
+    sys.exit();
+
+if not os.path.isfile(path_to_plot_exe):
+    print "{} does not exist!".format(path_to_plot_exe);
+    sys.exit();
 
 print 'Benchmarking OpenCL implementation with varying work group sizes.\n'
 
 tmp_file="tmp.txt"
 out_file="tmp_out.txt"
 log_file='benchmark_nbody_{}.log'.format(current_milli_time())
-
-work_group_sizes = [pow(2,x) for x in range(0,10)]
-
+work_group_sizes = [pow(2,x) for x in range(int(math.log(num_bodies,2))+1)]
 
 f = open(log_file, 'w');
 f.write('{}\n{}\n{}\n'.format(num_bodies,simulation_steps,num_repeats))
@@ -37,7 +52,6 @@ for work_group_size in work_group_sizes:
 
     print '\nBenchmarking for work_group_size={}\n'.format(work_group_size)
     for i in range(0,num_repeats):
-        # Benchmark OpenCl simulation
         simulation_call = '{} {} {} {} {} {}'.format(path_to_simulation_exe, tmp_file, simulation_steps, 'ocl', out_file, work_group_size)
         before = current_milli_time()
         os.system(simulation_call)
@@ -55,4 +69,4 @@ os.remove(tmp_file)
 os.remove(out_file)
 print 'Cleaned up temporary files.\n'
 
-os.system('./plot_benchmark_ocl.py {}'.format(log_file))
+os.system('{} {}'.format(path_to_plot_exe,log_file))
